+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.5.13
 
+* `Common`:
+    * Add functionality for multiplatform working with files:
+        * Main class for files `MPPFile`
+        * Inline class for filenames work encapsulation `FileName`
+
 ## 0.5.12
 
 * `Common`: